+++ Options/Applicative/Builder.hs
@@ -19,7 +19,6 @@ module Options.Applicative.Builder (
   -- creates a parser for an option called \"output\".
   subparser,
   argument,
-  argument',
   arguments,
   arguments1,
   flag,
@@ -63,6 +62,8 @@ module Options.Applicative.Builder (
   auto,
   str,
   disabled,
+  readerAbort,
+  readerError,
 
   -- * Builder for 'ParserInfo'
   InfoMod,
@@ -75,6 +76,7 @@ module Options.Applicative.Builder (
   progDesc,
   progDescDoc,
   failureCode,
+  noIntersperse,
   info,
 
   -- * Builder for 'ParserPrefs'
@@ -88,13 +90,14 @@ module Options.Applicative.Builder (
 
   -- * Types
   Mod,
+  ReadM,
   OptionFields,
   FlagFields,
   ArgumentFields,
   CommandFields
   ) where
 
-import Control.Applicative (pure, (<|>))
+import Control.Applicative (pure, (<|>), many, some)
 import Data.Monoid (Monoid (..)
 #if __GLASGOW_HASKELL__ > 702
   , (<>)
@@ -102,7 +105,6 @@ import Data.Monoid (Monoid (..)
   )
 
 import Options.Applicative.Builder.Completer
-import Options.Applicative.Builder.Arguments
 import Options.Applicative.Builder.Internal
 import Options.Applicative.Common
 import Options.Applicative.Types
@@ -209,6 +211,22 @@ subparser m = mkParser d g rdr
   where
     Mod _ d g = m `mappend` metavar "COMMAND"
     rdr = uncurry CmdReader (mkCommand m)
+
+-- | Builder for an argument parser.
+argument :: (String -> Maybe a) -> Mod ArgumentFields a -> Parser a
+argument p (Mod f d g) = mkParser d g (ArgReader rdr)
+  where
+    ArgumentFields compl = f (ArgumentFields mempty)
+    rdr = CReader compl p
+
+-- | Builder for an argument list parser. All arguments are collected and
+-- returned as a list.
+arguments :: (String -> Maybe a) -> Mod ArgumentFields a -> Parser [a]
+arguments r m = many (argument r m)
+
+-- | Like `arguments`, but require at least one argument.
+arguments1 :: (String -> Maybe a) -> Mod ArgumentFields a -> Parser [a]
+arguments1 r m = some (argument r m)
 
 -- | Builder for a flag parser.
 --
@@ -327,6 +345,10 @@ progDescDoc doc = InfoMod $ \i -> i { infoProgDesc = Chunk doc }
 failureCode :: Int -> InfoMod a
 failureCode n = InfoMod $ \i -> i { infoFailureCode = n }
 
+-- | Disable parsing of regular options after arguments
+noIntersperse :: InfoMod a
+noIntersperse = InfoMod $ \p -> p { infoIntersperse = False }
+
 -- | Create a 'ParserInfo' given a 'Parser' and a modifier.
 info :: Parser a -> InfoMod a -> ParserInfo a
 info parser m = applyInfoMod m base
@@ -337,7 +359,8 @@ info parser m = applyInfoMod m base
       , infoProgDesc = mempty
       , infoHeader = mempty
       , infoFooter = mempty
-      , infoFailureCode = 1 }
+      , infoFailureCode = 1
+      , infoIntersperse = True }
 
 newtype PrefsMod = PrefsMod
   { applyPrefsMod :: ParserPrefs -> ParserPrefs }